evaluaciones/Sumativa4/gestion_productos/requirements.txt
Django==5.1
django-ninja==1.1.0
djangorestframework==3.14.0
python-jose==3.3.0
pillow==10.2.0
python-decouple==3.8